fix(starryos): address K230 KPU review comments
