+++ test-suit/starryos/k230-qemu/qemu-k230/kpu-smoke/c/CMakeLists.txt
@@ -4,9 +4,6 @@ set(CMAKE_C_STANDARD 11)
 set(CMAKE_C_STANDARD_REQUIRED ON)
 set(CMAKE_C_EXTENSIONS OFF)
 add_executable(k230-kpu-smoke src/main.c)
-target_include_directories(k230-kpu-smoke PRIVATE
-    "${CMAKE_CURRENT_SOURCE_DIR}/../../../../../../drivers/npu/k230-kpu/include"
-)
 target_compile_options(k230-kpu-smoke PRIVATE -Wall -Wextra -Werror)
 install(TARGETS k230-kpu-smoke RUNTIME DESTINATION usr/bin)
 install(FILES
