+++ test-suit/starryos/k230-qemu/qemu-k230/kpu-smoke/c/src/main.c
@@ -14,7 +14,67 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
-#include "k230_kpu_uapi.h"
+#define KPU_DEVICE_PATH "/dev/kpu"
+#define KPU_DEVICE_ALIAS_PATH "/dev/kpu0"
+
+#define KPU_IOC_GET_STATUS 0x4b00u
+#define KPU_IOC_CLEAR 0x4b01u
+#define KPU_IOC_PROGRAM_COMMAND 0x4b02u
+#define KPU_IOC_START 0x4b03u
+#define KPU_IOC_RUN 0x4b04u
+#define KPU_IOC_WAIT_DONE 0x4b05u
+#define KPU_IOC_GET_INFO 0x4b06u
+#define KPU_IOC_GET_IRQ_COUNT 0x4b07u
+
+#define KPU_MMAP_CFG_OFFSET 0x0ull
+#define KPU_MMAP_L2_OFFSET 0x1000ull
+#define KPU_MMAP_FAKE_OUTPUT_OFFSET 0x2000ull
+#define KPU_MMAP_RUNTIME_RDATA_OFFSET 0x3000ull
+#define KPU_MMAP_RUNTIME_COMMAND_OFFSET 0x4000ull
+#define KPU_MMAP_RUNTIME_DIRECT_IO_OFFSET 0x5000ull
+#define KPU_MMAP_RUNTIME_DDR_OFFSET 0x6000ull
+
+#define KPU_CFG_PADDR 0x80400000ull
+#define KPU_L2_PADDR 0x80000000ull
+#define KPU_FAKE_OUTPUT_PADDR 0x10090000ull
+#define KPU_RUNTIME_RDATA_PADDR 0x10000000ull
+#define KPU_RUNTIME_COMMAND_PADDR 0x10190000ull
+#define KPU_RUNTIME_DIRECT_IO_PADDR 0x10500000ull
+#define KPU_RUNTIME_DDR_PADDR 0x3c000000ull
+
+#define KPU_RUNTIME_RDATA_BASE 0x10000020ull
+#define KPU_RUNTIME_FUNCTION_COMMAND_PADDR 0x1032b020ull
+#define KPU_RUNTIME_ARG_TABLE_PADDR 0x80000000ull
+#define KPU_RUNTIME_DIRECT_SOURCE_PADDR 0x10500020ull
+#define KPU_RUNTIME_DIRECT_OUTPUT_PADDR 0x10501020ull
+
+#define KPU_CFG_SIZE 0x800u
+#define KPU_L2_SIZE 0x200000u
+#define KPU_FAKE_OUTPUT_SIZE 0x100000u
+#define KPU_RUNTIME_RDATA_SIZE 0x90000u
+#define KPU_RUNTIME_COMMAND_SIZE 0x370000u
+#define KPU_RUNTIME_DIRECT_IO_SIZE 0xb00000u
+#define KPU_RUNTIME_DDR_SIZE 0x4000000u
+
+#define KPU_IRQ_NONE 0xffffffffu
+#define KPU_INFO_F_FDT 0x1u
+#define KPU_INFO_F_IRQ_WAIT 0x2u
+#define KPU_INFO_F_FAKE_OUTPUT 0x4u
+#define KPU_INFO_F_RUNTIME_SCRATCH 0x8u
+
+struct k230_kpu_command_range {
+    uint64_t start_paddr;
+    uint64_t end_paddr;
+};
+
+struct k230_kpu_info {
+    uint64_t cfg_paddr;
+    uint64_t cfg_size;
+    uint64_t l2_paddr;
+    uint64_t l2_size;
+    uint32_t irq;
+    uint32_t flags;
+};
 
 _Static_assert(sizeof(struct k230_kpu_command_range) == 16, "unexpected KPU command range size");
 _Static_assert(offsetof(struct k230_kpu_command_range, start_paddr) == 0,
